CurrencyRecognizer
This project detect the currency of bangladesh with 60%+ accuracy. If specific number of data is used as input, then we can get the 100% efficiency.

Opencv has been used to classify the raw data. OCR brings out the text in currency image(may not work all time). GUI has been created with TKinter.